- add test result - unify type

+++ rvp/builtin-rvp-smal.c
@@ -7,7 +7,7 @@
 
 #if defined(__riscv_zpsf) && !defined(__riscv_32e)
 static __attribute__ ((noinline))
-int64_t smal (int64_t ra, unsigned long rb)
+int64_t smal (int64_t ra, uint32_t rb)
 {
   return __rv__smal (ra, rb);
 }
