+++ rvp32_scan/builtin-rvp-smal.c
@@ -8,7 +8,7 @@
 #include <stdint.h>
 
 static __attribute__ ((noinline))
-int64_t lams (int64_t ra, unsigned long rb)
+int64_t lams (int64_t ra, uint32_t rb)
 {
   return __rv__smal (ra, rb);
 }
